Moved transaction related requests to contexts
