+++ bookkeeping-app/src/components/contexts/TransactionsCtx.jsx
@@ -4,8 +4,10 @@ import AuthCtx from "./AuthCtx";
 
 const TransactionsCtx = createContext({
     ctxTranList: null,
-    populateCtxTransactions: () => {},
     setCtxTranList: () => {},
+    populateCtxTransactions: () => {},
+    ctxAddTransactions: () => {},
+    ctxUpdateTransaction: () => {},
 });
 
 export function TransactionsCtxProvider(props) {
@@ -35,10 +37,80 @@ export function TransactionsCtxProvider(props) {
         }
     };
 
+    const ctxAddTransactions = async (transactionsToAdd) => {
+        const transformedTransactionsArray = transactionsToAdd.map((transaction) => ({
+            ...transaction,
+            entity_id: transaction.entity.id,
+            account_id: transaction.account.id,
+        }));
+
+        transformedTransactionsArray.forEach((transaction) => {
+            delete transaction.entity;
+            delete transaction.account;
+        });
+
+        try {
+            const response = await fetch("http://127.0.0.1:8000/api/transactions/", {
+                method: "POST",
+                headers: {
+                    "Content-Type": "application/json",
+                    Authorization: `Bearer ${ctxAccessToken}`,
+                },
+                body: JSON.stringify(transformedTransactionsArray),
+            });
+
+            if (!response.ok) {
+                console.log(response.error);
+                return;
+            }
+
+            const newData = await response.json();
+            setCtxTranList((prev) => [...prev, ...newData]);
+        } catch (error) {
+            console.error("Error sending transactions:", error);
+        }
+    };
+
+    const ctxUpdateTransaction = async (updatedTransaction) => {
+        let data;
+
+        try {
+            const response = await fetch(`http://127.0.0.1:8000/api/transactions/${updatedTransaction.id}/`, {
+                method: "PUT",
+                headers: {
+                    "Content-Type": "application/json",
+                    Authorization: `Bearer ${ctxAccessToken}`,
+                },
+                body: JSON.stringify(updatedTransaction),
+            });
+
+            if (!response.ok) {
+                console.log("Error:", response.error);
+                return;
+            }
+
+            const updatedData = await response.json();
+
+            setCtxTranList((prevTransactions) =>
+                prevTransactions.map((transaction) => {
+                    if (transaction.id === updatedTransaction.id) {
+                        return updatedData;
+                    } else {
+                        return transaction;
+                    }
+                })
+            );
+        } catch (error) {
+            console.error("Error editing transaction:", error);
+        }
+    };
+
     const context = {
         ctxTranList,
-        populateCtxTransactions,
         setCtxTranList,
+        populateCtxTransactions,
+        ctxAddTransactions,
+        ctxUpdateTransaction,
     };
 
     return <TransactionsCtx.Provider value={context}>{props.children}</TransactionsCtx.Provider>;
